isolates the PrividerConfOutput struct. integrates api manager data sources
