+++ anypoint/provider.go
@@ -3,30 +3,12 @@ package anypoint
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
+	"io"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
-	ame "github.com/mulesoft-anypoint/anypoint-client-go/ame"
-	ame_binding "github.com/mulesoft-anypoint/anypoint-client-go/ame_binding"
-	amq "github.com/mulesoft-anypoint/anypoint-client-go/amq"
 	auth "github.com/mulesoft-anypoint/anypoint-client-go/authorization"
-	connected_app "github.com/mulesoft-anypoint/anypoint-client-go/connected_app"
-	dlb "github.com/mulesoft-anypoint/anypoint-client-go/dlb"
-	env "github.com/mulesoft-anypoint/anypoint-client-go/env"
-	idp "github.com/mulesoft-anypoint/anypoint-client-go/idp"
-	org "github.com/mulesoft-anypoint/anypoint-client-go/org"
-	role "github.com/mulesoft-anypoint/anypoint-client-go/role"
-	rolegroup "github.com/mulesoft-anypoint/anypoint-client-go/rolegroup"
-	team "github.com/mulesoft-anypoint/anypoint-client-go/team"
-	team_group_mappings "github.com/mulesoft-anypoint/anypoint-client-go/team_group_mappings"
-	team_members "github.com/mulesoft-anypoint/anypoint-client-go/team_members"
-	team_roles "github.com/mulesoft-anypoint/anypoint-client-go/team_roles"
-	user "github.com/mulesoft-anypoint/anypoint-client-go/user"
-	user_rolegroups "github.com/mulesoft-anypoint/anypoint-client-go/user_rolegroups"
-	vpc "github.com/mulesoft-anypoint/anypoint-client-go/vpc"
-	vpn "github.com/mulesoft-anypoint/anypoint-client-go/vpn"
 )
 
 // Provider -
@@ -104,30 +86,33 @@ func Provider() *schema.Provider {
 			"anypoint_ame_binding":         resourceAMEBinding(),
 		},
 		DataSourcesMap: map[string]*schema.Resource{
-			"anypoint_vpcs":                dataSourceVPCs(),
-			"anypoint_vpc":                 dataSourceVPC(),
-			"anypoint_vpn":                 dataSourceVPN(),
-			"anypoint_bg":                  dataSourceBG(),
-			"anypoint_roles":               dataSourceRoles(),
-			"anypoint_rolegroup":           dataSourceRoleGroup(),
-			"anypoint_rolegroups":          dataSourceRoleGroups(),
-			"anypoint_users":               dataSourceUsers(),
-			"anypoint_user":                dataSourceUser(),
-			"anypoint_env":                 dataSourceENV(),
-			"anypoint_user_rolegroup":      dataSourceUserRolegroup(),
-			"anypoint_user_rolegroups":     dataSourceUserRolegroups(),
-			"anypoint_team":                dataSourceTeam(),
-			"anypoint_teams":               dataSourceTeams(),
-			"anypoint_team_roles":          dataSourceTeamRoles(),
-			"anypoint_team_members":        dataSourceTeamMembers(),
-			"anypoint_team_group_mappings": dataSourceTeamGroupMappings(),
-			"anypoint_dlb":                 dataSourceDLB(),
-			"anypoint_dlbs":                dataSourceDLBs(),
-			"anypoint_idp":                 dataSourceIDP(),
-			"anypoint_idps":                dataSourceIDPs(),
-			"anypoint_connected_app":       dataSourceConnectedApp(),
-			"anypoint_amq":                 dataSourceAMQ(),
-			"anypoint_ame":                 dataSourceAME(),
+			"anypoint_vpcs":                    dataSourceVPCs(),
+			"anypoint_vpc":                     dataSourceVPC(),
+			"anypoint_vpn":                     dataSourceVPN(),
+			"anypoint_bg":                      dataSourceBG(),
+			"anypoint_roles":                   dataSourceRoles(),
+			"anypoint_rolegroup":               dataSourceRoleGroup(),
+			"anypoint_rolegroups":              dataSourceRoleGroups(),
+			"anypoint_users":                   dataSourceUsers(),
+			"anypoint_user":                    dataSourceUser(),
+			"anypoint_env":                     dataSourceENV(),
+			"anypoint_user_rolegroup":          dataSourceUserRolegroup(),
+			"anypoint_user_rolegroups":         dataSourceUserRolegroups(),
+			"anypoint_team":                    dataSourceTeam(),
+			"anypoint_teams":                   dataSourceTeams(),
+			"anypoint_team_roles":              dataSourceTeamRoles(),
+			"anypoint_team_members":            dataSourceTeamMembers(),
+			"anypoint_team_group_mappings":     dataSourceTeamGroupMappings(),
+			"anypoint_dlb":                     dataSourceDLB(),
+			"anypoint_dlbs":                    dataSourceDLBs(),
+			"anypoint_idp":                     dataSourceIDP(),
+			"anypoint_idps":                    dataSourceIDPs(),
+			"anypoint_connected_app":           dataSourceConnectedApp(),
+			"anypoint_amq":                     dataSourceAMQ(),
+			"anypoint_ame":                     dataSourceAME(),
+			"anypoint_apim":                    dataSourceApim(),
+			"anypoint_apim_instance":           dataSourceApimInstance(),
+			"anypoint_apim_instance_upstreams": dataSourceApimInstanceUpstreams(),
 		},
 		ConfigureContextFunc: providerConfigure,
 		TerraformVersion:     "v1.0.1",
@@ -187,7 +172,7 @@ func userPwdAuth(ctx context.Context, username string, password string) (*auth.I
 	if err != nil {
 		var details string
 		if httpr != nil {
-			b, _ := ioutil.ReadAll(httpr.Body)
+			b, _ := io.ReadAll(httpr.Body)
 			details = string(b)
 		} else {
 			details = err.Error()
@@ -218,7 +203,7 @@ func connectedAppAuth(ctx context.Context, client_id string, client_secret strin
 	if err != nil {
 		var details string
 		if httpr != nil {
-			b, _ := ioutil.ReadAll(httpr.Body)
+			b, _ := io.ReadAll(httpr.Body)
 			details = string(b)
 		} else {
 			details = err.Error()
@@ -247,91 +232,4 @@ func cplane2serverindex(cplane string) int {
 		return 2
 	}
 	return -1
-}
-
-type ProviderConfOutput struct {
-	access_token            string
-	server_index            int
-	vpcclient               *vpc.APIClient
-	vpnclient               *vpn.APIClient
-	orgclient               *org.APIClient
-	roleclient              *role.APIClient
-	rolegroupclient         *rolegroup.APIClient
-	userclient              *user.APIClient
-	envclient               *env.APIClient
-	userrgpclient           *user_rolegroups.APIClient
-	teamclient              *team.APIClient
-	teammembersclient       *team_members.APIClient
-	teamrolesclient         *team_roles.APIClient
-	teamgroupmappingsclient *team_group_mappings.APIClient
-	dlbclient               *dlb.APIClient
-	idpclient               *idp.APIClient
-	connectedappclient      *connected_app.APIClient
-	amqclient               *amq.APIClient
-	ameclient               *ame.APIClient
-	amebindingclient        *ame_binding.APIClient
-}
-
-func newProviderConfOutput(access_token string, server_index int) ProviderConfOutput {
-	//preparing clients
-	vpccfg := vpc.NewConfiguration()
-	vpncfg := vpn.NewConfiguration()
-	orgcfg := org.NewConfiguration()
-	rolecfg := role.NewConfiguration()
-	rolegroupcfg := rolegroup.NewConfiguration()
-	usercfg := user.NewConfiguration()
-	envcfg := env.NewConfiguration()
-	userrolegroupscfg := user_rolegroups.NewConfiguration()
-	teamcfg := team.NewConfiguration()
-	teammemberscfg := team_members.NewConfiguration()
-	teamrolescfg := team_roles.NewConfiguration()
-	teamgroupmappingscfg := team_group_mappings.NewConfiguration()
-	dlbcfg := dlb.NewConfiguration()
-	idpcfg := idp.NewConfiguration()
-	connectedappcfg := connected_app.NewConfiguration()
-	amqcfg := amq.NewConfiguration()
-	amecfg := ame.NewConfiguration()
-	amebindingcfg := ame_binding.NewConfiguration()
-
-	vpcclient := vpc.NewAPIClient(vpccfg)
-	vpnclient := vpn.NewAPIClient(vpncfg)
-	orgclient := org.NewAPIClient(orgcfg)
-	roleclient := role.NewAPIClient(rolecfg)
-	rolegroupclient := rolegroup.NewAPIClient(rolegroupcfg)
-	userclient := user.NewAPIClient(usercfg)
-	envclient := env.NewAPIClient(envcfg)
-	userrgpclient := user_rolegroups.NewAPIClient(userrolegroupscfg)
-	teamclient := team.NewAPIClient(teamcfg)
-	teammembersclient := team_members.NewAPIClient(teammemberscfg)
-	teamrolesclient := team_roles.NewAPIClient(teamrolescfg)
-	teamgroupmappingsclient := team_group_mappings.NewAPIClient(teamgroupmappingscfg)
-	dlbclient := dlb.NewAPIClient(dlbcfg)
-	idpclient := idp.NewAPIClient(idpcfg)
-	connectedappclient := connected_app.NewAPIClient(connectedappcfg)
-	amqclient := amq.NewAPIClient(amqcfg)
-	ameclient := ame.NewAPIClient(amecfg)
-	amebindingclient := ame_binding.NewAPIClient(amebindingcfg)
-
-	return ProviderConfOutput{
-		access_token:            access_token,
-		server_index:            server_index,
-		vpcclient:               vpcclient,
-		vpnclient:               vpnclient,
-		orgclient:               orgclient,
-		roleclient:              roleclient,
-		rolegroupclient:         rolegroupclient,
-		userclient:              userclient,
-		envclient:               envclient,
-		userrgpclient:           userrgpclient,
-		teamclient:              teamclient,
-		teammembersclient:       teammembersclient,
-		teamrolesclient:         teamrolesclient,
-		teamgroupmappingsclient: teamgroupmappingsclient,
-		dlbclient:               dlbclient,
-		idpclient:               idpclient,
-		connectedappclient:      connectedappclient,
-		amqclient:               amqclient,
-		ameclient:               ameclient,
-		amebindingclient:        amebindingclient,
-	}
 }
